Created support for multiple notions of "changed"
Files Changed:
- DataImpression.cs
- DataImpressionGroup.cs
- SimpleGlyphDataImpression.cs
- SimpleLineDataImpression.cs
- SimpleSurfaceDataImpression.cs
- ABRState.cs

Because:
- Splitting up the "changed" flag into three finer-grain categories —
data changed, style changed, and visibility changed — improves the
performance of the engine by encouraging it to perform only the bare
minimum computations it needs to perform when a part of its state is
changed.
- Provides developers using the ABR engine as a Unity package better
control and performance as they incorporate the engine into their own
projects.

Further Explanation:
The previous "RenderHints.changed" flag has now been split into three
notions/levels: "RenderHints.DataChanged," "RenderHints.StyleChanged,"
and "RenderHints.VisibilityChanged." Additionally, the
"RenderHints.visible" flag has become "RenderHints.Visible" (NOTE: this
has yet to be recorded in the ABR json schema). These changes required
the addition of two new methods in the IDataImpression interface:
UpdateStyling() and SetVisibility(). SetVisibility() is the simplest,
working by directly toggling the meshrenderer components of impression
gameobjects in accordance with the visibility boolean passed.
UpdateStyling() is considerably more complex and varies for each type
of impression,  but in short it works by obtaining the most recent
"styling" information (scalars, colormaps, textures, etc.) and applying
it to the material used by the impression. In order to bring this all
together one further functionality had to be implemented in the ABRState
to actually determine which flags to enable/disable. This functionality
exists as a block of code in the middle of ABRState's LoadState()
method, and works by generating a RawABRState of the previous json
state and comparing the input values in that state to those in the
current state to determine which inputs have changed, and therefore
which flags to toggle.